cli: add --headers_only option to notmuch-show.c
This option just outputs the headers of a message (i.e. it omits the
bodies). Currently this is only implemented for format=json.

This is used by notmuch-pick.el (although not needed) because it gives a
speed-up of at least a factor of a two; moreover it reduces the memory
usage in emacs hugely.